adding authentication requirements to server routes

+++ server/routes/events.router.js
@@ -4,11 +4,12 @@ const router = express.Router();
 
 //GET Route
 router.get('/', (req, res) => {
-  /* We are breaking apart our GET call to the Database 
-  so that the date comes formatted in Month Date and Year, 
-  otherwise this will pull with a weird time-stamp. */
-  const sqlText =
-    `SELECT 
+  if (req.isAuthenticated()) {
+    /* We are breaking apart our GET call to the Database 
+    so that the date comes formatted in Month Date and Year, 
+    otherwise this will pull with a weird time-stamp. */
+    const sqlText =
+      `SELECT 
     "id", 
     "event_name",
     TO_CHAR("date", 'MM-DD-YYYY') AS formatted_date,
@@ -17,17 +18,18 @@ router.get('/', (req, res) => {
     "notes",
     "event_complete"
     FROM events ORDER BY "date" ASC`;
-  // console.log("Get text", sqlText);
-  pool
-    .query(sqlText)
-    .then((result) => {
-      // console.log(`GET from database`, result);
-      res.send(result.rows);
-    })
-    .catch((error) => {
-      console.log(`Error making database query ${sqlText}`, error);
-      res.sendStatus(500);
-    });
+    // console.log("Get text", sqlText);
+    pool
+      .query(sqlText)
+      .then((result) => {
+        // console.log(`GET from database`, result);
+        res.send(result.rows);
+      })
+      .catch((error) => {
+        console.log(`Error making database query ${sqlText}`, error);
+        res.sendStatus(500);
+      });
+  }
 });
 
 //POST Route
@@ -49,11 +51,12 @@ router.post('/', (req, res) => {
 
 // Update Event Data with a person id
 router.put('/:eventId', (req, res) => {
-  let idToUpdate = req.body.eventID
-  console.log("Id to Update:", idToUpdate)
-  let eventData = req.body;
-  console.log("Event Data is:", eventData)
-  let sqlText = `UPDATE events 
+  if (req.isAuthenticated()) {
+    let idToUpdate = req.body.eventID
+    console.log("Id to Update:", idToUpdate)
+    let eventData = req.body;
+    console.log("Event Data is:", eventData)
+    let sqlText = `UPDATE events 
   SET 
   "event_name" = $2,
   "date" = $3,
@@ -62,35 +65,38 @@ router.put('/:eventId', (req, res) => {
   "notes" = $6 
   WHERE "id" = $1;`;
 
-  pool
-    .query(sqlText, [idToUpdate, eventData.eventName, eventData.date, eventData.time, eventData.address, eventData.notes])
-    .then((result) => {
-      console.log("ID updated in database", idToUpdate);
-      res.sendStatus(200);
-    })
-    .catch((error) => {
-      console.log(`Error making database query ${sqlText}`, error);
-      res.sendStatus(500);
-    });
+    pool
+      .query(sqlText, [idToUpdate, eventData.eventName, eventData.date, eventData.time, eventData.address, eventData.notes])
+      .then((result) => {
+        console.log("ID updated in database", idToUpdate);
+        res.sendStatus(200);
+      })
+      .catch((error) => {
+        console.log(`Error making database query ${sqlText}`, error);
+        res.sendStatus(500);
+      });
+  }
 });
 
 
 router.delete('/:id', (req, res) => {
-  let idToDelete = req.params.id;
-  console.log("idToDelete", idToDelete);
-  let sqlText = `
+  if (req.isAuthenticated()) {
+    let idToDelete = req.params.id;
+    console.log("idToDelete", idToDelete);
+    let sqlText = `
         DELETE FROM events WHERE "id" = $1;
         `;
-  pool
-    .query(sqlText, [idToDelete])
-    .then((result) => {
-      console.log("Deleted from database ", idToDelete);
-      res.sendStatus(202);
-    })
-    .catch((error) => {
-      console.log(`Error making database query ${sqlText}`, error);
-      res.sendStatus(500);
-    });
+    pool
+      .query(sqlText, [idToDelete])
+      .then((result) => {
+        console.log("Deleted from database ", idToDelete);
+        res.sendStatus(202);
+      })
+      .catch((error) => {
+        console.log(`Error making database query ${sqlText}`, error);
+        res.sendStatus(500);
+      });
+  }
 });
 
 module.exports = router;
